# Optimize Ledger.Get() by making Forest.Read() use ~5x fewer allocs/op and run ~20% faster