--- a/ledger/complete/mtrie/forest.go
+++ b/ledger/complete/mtrie/forest.go
@@ -123,8 +123,8 @@ func (f *Forest) ValueSizes(r *ledger.TrieRead) ([]int, error) {
 	return orderedValueSizes, nil
 }
 
-// ReadSinglePayload reads value for a single path and returns value and error (if any)
-func (f *Forest) ReadSinglePayload(r *ledger.TrieReadSinglePayload) (*ledger.Payload, error) {
+// ReadSingleValue reads value for a single path and returns value and error (if any)
+func (f *Forest) ReadSingleValue(r *ledger.TrieReadSingleValue) (ledger.Value, error) {
 	// lookup the trie by rootHash
 	trie, err := f.GetTrie(r.RootHash)
 	if err != nil {
@@ -132,7 +132,7 @@ func (f *Forest) ReadSinglePayload(r *ledger.TrieReadSinglePayload) (*ledger.Pay
 	}
 
 	payload := trie.ReadSinglePayload(r.Path)
-	return payload.DeepCopy(), nil
+	return payload.Value.DeepCopy(), nil
 }
 
 // Read reads values for an slice of paths and returns values and error (if any)
@@ -152,7 +152,7 @@ func (f *Forest) Read(r *ledger.TrieRead) ([]ledger.Value, error) {
 	// call ReadSinglePayload if there is only one path
 	if len(r.Paths) == 1 {
 		payload := trie.ReadSinglePayload(r.Paths[0])
-		return []*ledger.Payload{payload.DeepCopy()}, nil
+		return []ledger.Value{payload.Value.DeepCopy()}, nil
 	}
 
 	// deduplicate keys: