### Changes

Change `Forest.Read()` and `Forest.ReadSingleValue()` to return `[]ledger.Value` without deep copying payload keys.  This avoids 4 heap allocation per key.

This change doesn't affect `Ledger.Get()` (the caller) because it discards the payload keys.

Closes #2475 

### Benchmark Comparisons (includes PR #2473 + this PR)

Bench comparison of master `Ledger.Get()` vs `Ledger.GetSingleValue()` reading a single value.

```
name          old time/op    new time/op    delta
LedgerGet1-4    6.77µs ± 1%    4.30µs ± 1%  -36.47%  (p=0.000 n=9+9)

name          old alloc/op   new alloc/op   delta
LedgerGet1-4    1.74kB ± 0%    0.69kB ± 0%  -60.55%  (p=0.000 n=10+10)

name          old allocs/op  new allocs/op  delta
LedgerGet1-4      21.0 ± 0%       3.0 ± 0%  -85.71%  (p=0.000 n=10+10)
```

Bench comparison of master `Ledger.Get()` vs new `Ledger.Get()` reading 100 values.
```
name            old time/op    new time/op    delta
LedgerGet100-4     519µs ± 0%     410µs ± 1%  -21.02%  (p=0.000 n=10+9)

name            old alloc/op   new alloc/op   delta
LedgerGet100-4     190kB ± 0%      95kB ± 0%  -50.04%  (p=0.000 n=10+10)

name            old allocs/op  new allocs/op  delta
LedgerGet100-4     1.52k ± 0%     0.32k ± 0%  -79.17%  (p=0.000 n=10+10)
```

Benchmarks used Go 1.17 on linux_amd64 (Haswell).