@fxamacker 2022-05-25
@tarakby Thanks for taking a look and great questions!

> func (f *Forest) Read calls func (mt *MTrie) UnsafeRead which returns []*Payload, it then takes only the Value part. I was wondering if we could make UnsafeRead return []Value instead. We won't be able to track totalPayloadSize anymore, we would be able to track totalValueSize instead (is that a problem?).

The main optimization is to avoid payload key deep copy which happens in `Forest.Read`.  For keys containing owner, controller, and key, there are 4 allocs per copy because of its data structure. 

It probably won't  give us much speedup or reduce allocs if we make `UnsafeRead` return `[]Value` instead of `[]*Payload` because:
- `UnsafeRead()` doesn't deep copy returned payloads, so there isn't extra allocs even if it returns `[]Value`.
- `Forest.Read()` needs to create new `[]ledger.Value` to ensure values are in the same order as received paths.  So this alloc needs to happen even if `UnsafeRead()` returns `[]Value`.

If we only return `[]Value` from `MTrie.UnsafeRead`, callers won't have access to payload key anymore.  The payload key could be useful for callers.

I don't know if it would be a problem to replace `totalPayloadSize` with `totalValueSize` in metrics reports.  I can find out from @ramtinms if needed.

> UnsafeRead seems to handle the case of len(paths)==1 separately to avoid the recursive overhead. Do you think there is still advantage in handling that edge case inside Read ?

Great question!  Yes because they optimize different edge cases.

In [`MTrie.UnsafeRead()`](https://github.com/onflow/flow-go/blob/31d144c4a8727961f030e6743c1e587d7685966c/ledger/complete/mtrie/trie/trie.go#L271-L276), `len(paths) == 1` optimizes read when there is only one path to read in subtree as result of partitioning.     

In [`Forest.Read()`](https://github.com/onflow/flow-go/blob/31d144c4a8727961f030e6743c1e587d7685966c/ledger/complete/mtrie/forest.go#L152-L156), `len(paths) == 1` optimizes read when there is only one path to batch read for entire mtrie.  It avoids the overhead of deduplicating paths and reconstructing values in order after read.